+++ src/cdb/binary.rs
@@ -1,5 +1,4 @@
 use {
-    super::ClauseId,
     crate::types::*,
     std::{
         collections::HashMap,
@@ -8,7 +7,7 @@ use {
 };
 
 /// storage of binary links
-pub type BinaryLinkList = Vec<(Lit, ClauseId)>;
+pub type BinaryLinkList = Vec<(Lit, ClauseIndex)>;
 
 impl Index<Lit> for Vec<BinaryLinkList> {
     type Output = BinaryLinkList;
@@ -35,11 +34,11 @@ impl IndexMut<Lit> for Vec<BinaryLinkList> {
     }
 }
 
-/// storage with mapper to `ClauseId` of binary links
+/// storage with mapper to `ClauseIndex` of binary links
 #[derive(Clone, Debug, Default)]
 pub struct BinaryLinkDB {
-    hash: HashMap<(Lit, Lit), ClauseId>,
-    list: Vec<BinaryLinkList>,
+    pub(super) hash: HashMap<(Lit, Lit), ClauseIndex>,
+    pub(super) list: Vec<BinaryLinkList>,
 }
 
 impl Instantiate for BinaryLinkDB {
@@ -54,12 +53,12 @@ impl Instantiate for BinaryLinkDB {
 }
 
 pub trait BinaryLinkIF {
-    /// add a mapping from a pair of Lit to a `ClauseId`
-    fn add(&mut self, lit0: Lit, lit1: Lit, cid: ClauseId);
+    /// add a mapping from a pair of Lit to a `ClauseIndex`
+    fn add(&mut self, lit0: Lit, lit1: Lit, cid: ClauseIndex);
     /// remove a pair of `Lit`s
     fn remove(&mut self, lit0: Lit, lit1: Lit) -> MaybeInconsistent;
-    /// return 'ClauseId` linked from a pair of `Lit`s
-    fn search(&self, lit0: Lit, lit1: Lit) -> Option<&ClauseId>;
+    /// return 'ClauseIndex` linked from a pair of `Lit`s
+    fn search(&self, lit0: Lit, lit1: Lit) -> Option<&ClauseIndex>;
     /// return the all links that include `Lit`.
     /// Note this is not a `watch_list`. The other literal has an opposite phase.
     fn connect_with(&self, lit: Lit) -> &BinaryLinkList;
@@ -70,7 +69,7 @@ pub trait BinaryLinkIF {
 }
 
 impl BinaryLinkIF for BinaryLinkDB {
-    fn add(&mut self, lit0: Lit, lit1: Lit, cid: ClauseId) {
+    fn add(&mut self, lit0: Lit, lit1: Lit, cid: ClauseIndex) {
         let l0 = lit0.min(lit1);
         let l1 = lit0.max(lit1);
         self.hash.insert((l0, l1), cid);
@@ -85,7 +84,7 @@ impl BinaryLinkIF for BinaryLinkDB {
         self.list[lit1].delete_unstable(|p| p.0 == lit0);
         Ok(())
     }
-    fn search(&self, lit0: Lit, lit1: Lit) -> Option<&ClauseId> {
+    fn search(&self, lit0: Lit, lit1: Lit) -> Option<&ClauseIndex> {
         let l0 = lit0.min(lit1);
         let l1 = lit0.max(lit1);
         self.hash.get(&(l0, l1))
